requirements.txt
torch
torchvision
torchaudio
transformers
flash-attn
soundfile
librosa
jsonlines
gradio
pyannote.audio
